split message_dispatcher_service -> block_service, message_service
