+++ txpool/block_service.go
@@ -1,0 +1,5 @@
+package txpool
+
+type BlockService interface {
+	ProposeBlock(transactions []Transaction) error
+}
