+++ txpool/infra/adapter/command_handler.go
@@ -1,4 +1,4 @@
-package messaging
+package adapter
 
 import (
 	"log"
