+++ txpool/infra/message_service_test.go
@@ -1,4 +1,4 @@
-package messaging
+package infra
 
 import (
 	"errors"
